d/aws_billing_service_account: Move to 'billing' service package.
