+++ internal/service/billing/billing_service_account_data_source.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package meta
+package billing
 
 import (
 	"context"
